# ACM-18288: optimize the agent reconciliation rate