… hack/crds/hypershift.openshift.io_hostedclusters.yaml